backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
google-genai==0.3.2
manim==0.18.0
pillow==10.1.0
aiofiles==23.2.1
python-json-logger==2.0.7
tenacity==8.2.3
pydantic-settings==2.0.3 